+++ docs/gitbook/rust/changelog.md
@@ -1,3 +1,17 @@
+## [1.2.8](https://github.com/taskforcesh/bullmq/compare/vrs1.2.7...vrs1.2.8) (2026-08-30)
+
+
+### Bug Fixes
+
+* **deps:** upgrade msgpackr to 2.1.0 and uuid to 1.26.0 (rust) ([#4621](https://github.com/taskforcesh/bullmq/issues/4621)) ([74987e9](https://github.com/taskforcesh/bullmq/commit/74987e914f22e7647662406861e2ad3101b883d4))
+
+## [1.2.7](https://github.com/taskforcesh/bullmq/compare/vrs1.2.6...vrs1.2.7) (2026-08-25)
+
+
+### Performance Improvements
+
+* **queue:** do not affect rate limit when processing deferred failures (python) (elixir) (rust) (dotnet) ([#4607](https://github.com/taskforcesh/bullmq/issues/4607)) ([ec4be04](https://github.com/taskforcesh/bullmq/commit/ec4be04c04725f5e341920ddef98adf7874a41d5))
+
 ## [1.2.6](https://github.com/taskforcesh/bullmq/compare/vrs1.2.5...vrs1.2.6) (2026-08-23)